Moved logic to general seekbar-overlay
