+++ web/template/watch.gohtml
@@ -230,32 +230,46 @@
             {{end}}
 
             <div
-                    x-data="{ size: { x: '0px', width: '0px' }, seekLogger: new watch.SeekLogger('{{$stream.ID}}'), heatMapPath: null, seekbarMarker: [] }"
-                    x-on:updateheatmappath.window="e => { heatMapPath = e.detail.heatMapPath; }"
+                    x-data="{ size: { x: '0px', width: '0px' } }"
                     x-on:updateseekbarsize.window="e => { size = e.detail; }"
-                    x-on:seekbarhighlightsupdate.window="e => { seekbarMarker = e.detail.marker; seekbarSections = e.detail.sections; }"
-                    x-init="seekLogger.attach(); watch.repeatHeatMap.init('{{$stream.ID}}'); watch.seekbarHighlights.init('{{$stream.ID}}');"
-                    class="heatmap-wrap absolute bottom-[18px] rounded bg-gradient-to-t from-gray-800/[.6] pt-4 pointer-events-none"
+                    x-init="watch.seekbarOverlay.init('{{$stream.ID}}');"
+                    class="seekbar-overlay absolute bottom-[18px] bg-gradient-to-t from-gray-800/[.6] pt-4 pointer-events-none"
                     :style="{width: size.width, left: size.x}"
-                    x-show="heatMapPath != null"
             >
-                <template x-if="heatMapPath">
-                    <svg height="100%" preserveAspectRatio="none" viewBox="0 0 1000 100"
-                         width="100%" style="height: 30px;">
-                        <defs>
-                            <clipPath id="heatmap">
-                                <path class="tumlive-heat-map-path"
-                                      :d="heatMapPath"
-                                      fill="white" fill-opacity="0.6"></path>
-                            </clipPath>
-                        </defs>
-                        <rect clip-path="url(#heatmap)" fill="white" fill-opacity="0.4" height="100%" width="100%" x="0" y="0"></rect>
-                    </svg>
-                </template>
 
-                <template>
+                <div
+                        x-data="{ seekLogger: new watch.SeekLogger('{{$stream.ID}}'), heatMapPath: null }"
+                        x-on:updateheatmappath.window="e => { heatMapPath = e.detail.heatMapPath; }"
+                        x-on:updateseekbarsize.window="e => { size = e.detail; }"
+                        x-init="seekLogger.attach(); watch.repeatHeatMap.init('{{$stream.ID}}');"
+                        class="absolute h-[30px] bottom-0 left-0 right-0"
+                        x-show="heatMapPath != null"
+                >
+                    <template x-if="heatMapPath">
+                        <svg height="100%" preserveAspectRatio="none" viewBox="0 0 1000 100"
+                             width="100%" style="height: 30px;">
+                            <defs>
+                                <clipPath id="heatmap">
+                                    <path class="tumlive-heat-map-path"
+                                          :d="heatMapPath"
+                                          fill="white" fill-opacity="0.6"></path>
+                                </clipPath>
+                            </defs>
+                            <rect clip-path="url(#heatmap)" fill="white" fill-opacity="0.4" height="100%" width="100%" x="0" y="0"></rect>
+                        </svg>
+                    </template>
+                </div>
 
-                </template>
+                <div
+                        x-data="{ seekbarMarker: [], seekbarSections: [] }"
+                        class="w-[100%] h-[30px] relative"
+                        x-on:seekbarhighlightsupdate.window="e => { seekbarMarker = e.detail.marker; seekbarSections = e.detail.sections; }"
+                        x-init="watch.seekbarHighlights.init('{{$stream.ID}}');"
+                >
+                    <template x-for="marker in seekbarMarker">
+                        <div class="absolute bg-white bottom-[18px] top-[0] w-[2px]" :style="{ left: marker.pos }"></div>
+                    </template>
+                </div>
             </div>
 
             <div
